// src/app/core/enums/skills/skillsNameEnum.ts
/**
 * @description 定義 Skill 的名稱
 */
export enum SkillsNameEnum {
  JAVASCRIPT = 'JavaScript',
  HTML = 'HTML',
  CSS = 'CSS',
  VUE = 'Vue.js',
  REACT = 'React',
  NODE = 'Node.js',
  EXPRESS = 'Express',
  MONGODB = 'MongoDB',
  SQL = 'SQL',
  GIT = 'Git',
  WEBPACK = 'Webpack',
  BABEL = 'Babel',
  TYPESCRIPT = 'TypeScript',
  GRAPHQL = 'GraphQL',
  SASS = 'Sass',
  BOOTSTRAP = 'Bootstrap',
  DOCKER = 'Docker',
  KUBERNETES = 'Kubernetes',
  AWS = 'AWS',
  AZURE = 'Azure',
  FIREBASE = 'Firebase',
  JAVA = 'Java',
  PYTHON = 'Python',
  CPP = 'C++',
  RUBY = 'Ruby',
  PHOTOSHOP = 'Photoshop',
  ILLUSTRATOR = 'Illustrator',
  SKETCH = 'Sketch',
  FIGMA = 'Figma',
  WIRE_FRAMING = 'Wire Framing',
  PROTOTYPING = 'Prototyping',
  PROJECT_MANAGEMENT = 'Project Management',
  SCRUM = 'Scrum',
  AGILE = 'Agile',
  JIRA = 'JIRA',
  MARKET_RESEARCH = 'Market Research',
  SEO = 'SEO',
  CONTENT_MARKETING = 'Content Marketing',
  SOCIAL_MEDIA_MARKETING = 'Social Media Marketing',
  GOOGLE_ANALYTICS = 'Google Analytics',
  ACCOUNTING = 'Accounting',
  FINANCIAL_ANALYSIS = 'Financial Analysis',
  INVESTMENT_STRATEGIES = 'Investment Strategies',
  RISK_MANAGEMENT = 'Risk Management',
  BUDGETING = 'Budgeting',
  COMMUNICATION = 'Communication',
  TEAMWORK = 'Teamwork',
  PROBLEM_SOLVING = 'Problem Solving',
  LEADERSHIP = 'Leadership',
  TIME_MANAGEMENT = 'Time Management',
  ADAPTABILITY = 'Adaptability',
  CREATIVITY = 'Creativity',
  WORK_ETHIC = 'Work Ethic',
  INTERPERSONAL_SKILLS = 'Interpersonal Skills'
}